Rename component Canvas -> Tile
